8952 Move formats tests
